Refactor page titles and add GitHub login option
