+++ frontend/src/pages/auth/login/LoginScreen.tsx
@@ -2,52 +2,55 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import { Github } from "lucide-react";
 import { Link } from "react-router-dom";
 
 const LoginScreen = () => {
     return (
-        <Card className="mx-auto max-w-sm">
-            <CardHeader>
-                <CardTitle className="text-2xl">Login</CardTitle>
-                <CardDescription>
-                    Enter your email below to login to your account
-                </CardDescription>
-            </CardHeader>
-            <CardContent>
-                <div className="grid gap-4">
-                    <div className="grid gap-2">
-                        <Label htmlFor="email">Email</Label>
-                        <Input
-                            id="email"
-                            type="email"
-                            placeholder="m@example.com"
-                            required
-                        />
-                    </div>
-                    <div className="grid gap-2">
-                        <div className="flex items-center">
-                            <Label htmlFor="password">Password</Label>
-                            <Link to={"#"} className="ml-auto inline-block text-sm underline">
-                                Forgot your password?
-                            </Link>
+        <main className="h-screen grid place-items-center">
+            <Card className="mx-auto max-w-sm">
+                <CardHeader>
+                    <CardTitle className="text-2xl">Login</CardTitle>
+                    <CardDescription>
+                        Enter your email below to login to your account
+                    </CardDescription>
+                </CardHeader>
+                <CardContent>
+                    <div className="grid gap-4">
+                        <div className="grid gap-2">
+                            <Label htmlFor="email">Email</Label>
+                            <Input
+                                id="email"
+                                type="email"
+                                placeholder="m@example.com"
+                                required
+                            />
                         </div>
-                        <Input id="password" type="password" required />
+                        <div className="grid gap-2">
+                            <div className="flex items-center">
+                                <Label htmlFor="password">Password</Label>
+                                <Link to={"#"} className="ml-auto inline-block text-sm underline">
+                                    Forgot your password?
+                                </Link>
+                            </div>
+                            <Input id="password" type="password" required />
+                        </div>
+                        <Button type="submit" className="w-full">
+                            Login
+                        </Button>
+                        <Button variant="outline" className="w-full">
+                            <Github size={24} />
+                        </Button>
                     </div>
-                    <Button type="submit" className="w-full">
-                        Login
-                    </Button>
-                    <Button variant="outline" className="w-full">
-                        Login with Google
-                    </Button>
-                </div>
-                <div className="mt-4 text-center text-sm">
-                    Don&apos;t have an account?{" "}
-                    <Link to={"/auth/register"} className="underline">
-                        Sign up
-                    </Link>
-                </div>
-            </CardContent>
-        </Card>
+                    <div className="mt-4 text-center text-sm">
+                        Don&apos;t have an account?{" "}
+                        <Link to={"/auth/register"} className="underline">
+                            Sign up
+                        </Link>
+                    </div>
+                </CardContent>
+            </Card>
+        </main>
     )
 }
