+++ frontend/src/pages/auth/register/RegisterScreen.tsx
@@ -1,10 +1,11 @@
 import { useState } from "react";
-import { registerUser } from "./assets/utils/handleSubmit/handleSubmit";
+import { registerUser } from "./assets/utils/handleSubmit/registerUser";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Link } from "react-router-dom";
+import { Github } from "lucide-react";
 
 interface FormData {
     username: string;
@@ -37,7 +38,7 @@ const RegisterScreen = () => {
     };
 
     return (
-        <>
+        <main className="h-screen grid place-items-center">
             <Card className="mx-auto max-w-sm">
                 <CardHeader>
                     <CardTitle className="text-xl">Sign Up</CardTitle>
@@ -92,7 +93,7 @@ const RegisterScreen = () => {
                                 Create an account
                             </Button>
                             <Button variant="outline" className="w-full">
-                                Sign up with GitHub
+                                <Github size={24} />
                             </Button>
                         </div>
                         <div className="mt-4 text-center text-sm">
@@ -104,7 +105,7 @@ const RegisterScreen = () => {
                     </form>
                 </CardContent>
             </Card>
-        </>
+        </main>
     )
 }
