+++ frontend/src/pages/auth/register/assets/utils/handleSubmit/registerUser.tsx
